requirement.txt
fastapi
uvicorn
python-dateutil
starlette
mysql-connector-python
pydantic
